+++ Admin/MenuNodeAdmin.php
@@ -8,13 +8,12 @@ use Sonata\AdminBundle\Show\ShowMapper;
 use Sonata\DoctrinePHPCRAdminBundle\Admin\Admin;
 use Symfony\Cmf\Bundle\MenuBundle\Document\MenuNode;
 use Symfony\Component\HttpFoundation\Request;
+use Knp\Menu\ItemInterface as MenuItemInterface;
+use Symfony\Cmf\Bundle\MenuBundle\ContentAwareFactory;
+use Doctrine\Common\Util\ClassUtils;
 
-class MenuNodeAdmin extends Admin
+class MenuNodeAdmin extends MenuAdmin
 {
-    protected $translationDomain = 'CmfMenuBundle';
-    protected $contentRoot;
-    protected $menuRoot;
-
     protected function configureListFields(ListMapper $listMapper)
     {
         $listMapper
@@ -45,7 +44,22 @@ class MenuNodeAdmin extends Admin
 
         if (null === $this->getParentFieldDescription()) {
             $formMapper
-                ->with('form.group_general')
+                ->with('form.group_target', array(
+                    'template' => 'CmfMenuBundle:Admin:menu_node_target_group.html.twig',
+                ))
+                ->add('linkType', 'choice_field_mask', array(
+                    'choices' => array_combine(
+                        $this->contentAwareFactory->getLinkTypes(),
+                        $this->contentAwareFactory->getLinkTypes()
+                    ),
+                    'map' => array(
+                        'route' => array('route'),
+                        'uri' => array('uri'),
+                        'content' => array('content', 'doctrine_phpcr_odm_tree', 'weak'),
+                    ),
+                    'empty_value' => 'auto',
+                ))
+                ->add('weak', 'checkbox', array('required' => false))
                 ->add('route', 'text', array('required' => false))
                 ->add('uri', 'text', array('required' => false))
                 ->add(
@@ -124,5 +138,84 @@ class MenuNodeAdmin extends Admin
                 )
             );
         }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function buildBreadcrumbs($action, MenuItemInterface $menu = null)
+    {
+        $menuNodeNode = parent::buildBreadcrumbs($action, $menu);
+
+        if ($action != 'edit') {
+            return $menuNodeNode;
+        }
+
+        $menuDoc = $this->getMenuForSubject($this->getSubject());
+        $pool = $this->getConfigurationPool();
+        $menuAdmin = $pool->getAdminByClass(
+            ClassUtils::getClass($menuDoc)
+        );
+        $menuAdmin->setSubject($menuDoc);
+        $menuEditNode = $menuAdmin->buildBreadcrumbs($action, $menu);
+        if ($menuAdmin->isGranted('EDIT' && $menuAdmin->hasRoute('edit'))) {
+            $menuEditNode->setUri(
+                $menuAdmin->generateUrl('edit', array(
+                    'id' => $this->getUrlsafeIdentifier($menuDoc)
+                ))
+            );
+        }
+
+        $menuNodeNode->setParent(null);
+        $current = $menuEditNode->addChild($menuNodeNode);
+
+        return $current;
+    }
+
+    protected function getMenuForSubject(MenuNode $subject)
+    {
+        $id = $subject->getId();
+
+        $menuId = $this->getMenuIdForNodeId($id);
+
+        $menu = $this->modelManager->find(null, $menuId);
+
+        return $menu;
+    }
+
+    protected function getMenuIdForNodeId($id)
+    {
+        // I wonder if this could be simplified in Phpcr/PathHelper
+        //
+        // $relPath = PathHelper::removeBasePath($this->menuRoot, $id);
+        // $menuId = PathHelper:splicePath($relPath, 0, 1);
+
+        if (0 !== strpos($id, $this->menuRoot)) {
+            throw new \InvalidArgumentException(sprintf(
+                'Cannot find base path "%s" in menu node ID "%s"', $this->menuRoot, $id
+            ));
+        }
+
+        $relPath = substr($id, strlen($this->menuRoot) + 1);
+        $parts = explode('/', $relPath);
+
+        if (count($parts) == 0) {
+            throw new \InvalidArgumentException(sprintf(
+                'ID for menu node "%s" is the same as root path "%s" - this is strange.',
+                $id, $rootPath
+            ));
+        }
+
+        if (count($parts) == 1) {
+            throw new \InvalidArgumentException(sprintf(
+                'MenuNode "%s" seems to hold the position reserved for a Menu. This should not happen',
+                $id
+            ));
+        }
+
+        $first = $parts[0];
+        $menuId = sprintf('%s/%s', $this->menuRoot, $first);
+
+        return $menuId;
     }
 }