+++ Document/MenuNode.php
@@ -15,26 +15,36 @@ use Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow\PublishTimePeriodWriteInterfac
 class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, PublishableWriteInterface
 {
     /**
-     * Id of this menu node
+     * Id of this menu node.
+     *
      * @var string
      */
     protected $id;
 
     /**
-     * Parent node
+     * Parent node.
+     *
      * @var mixed
      */
     protected $parent;
 
     /**
-     * Node name
+     * Node name.
+     *
      * @var string
      */
     protected $name;
 
+    /**
+     * Child menu nodes
+     *
+     * @var MenuNode[]
+     */
+    protected $children = array();
 
     /**
-     * Menu label
+     * Menu label.
+     *
      * @var string
      */
     protected $label = '';
@@ -45,9 +55,17 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     protected $uri;
 
     /**
+     * The name of the route to generate.
+     *
      * @var string
      */
     protected $route;
+
+    /**
+     * Enum, values determined by ContentAwareFactory
+     * @var string
+     */
+    protected $linkType;
 
     /**
      * @var mixed
@@ -60,37 +78,86 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     protected $hardContent;
 
     /**
-     * If we should use the weak or the strong
-     * referened content.
+     * If we should use the weak or the strong referenced content.
+     *
      * @var boolean
      */
     protected $weak = true;
 
     /**
+     * HTML attributes to add to the individual menu element.
+     *
+     * e.g. array('class' => 'foobar', 'style' => 'bar: foo')
+     *
      * @var array
      */
     protected $attributes = array();
 
     /**
+     * HTML attributes to add to the children list element.
+     *
+     * e.g. array('class' => 'foobar', 'style' => 'bar: foo')
+     *
      * @var array
      */
     protected $childrenAttributes = array();
 
     /**
+     * HTML attributes to add to items link.
+     *
+     * e.g. array('class' => 'foobar', 'style' => 'bar: foo')
+     *
      * @var array
      */
-    protected $children = array();
+    protected $linkAttributes = array();
 
     /**
-     * Hashmap for extra stuff associated to the node
+     * HTML attributes to add to the items label.
+     *
+     * e.g. array('class' => 'foobar', 'style' => 'bar: foo')
+     *
+     * @var array
+     */
+    protected $labelAttributes = array();
+
+    /**
+     * Hashmap for extra stuff associated to the node.
+     *
      * @var array
      */
     protected $extras;
 
     /**
+     * Parameters to use when generating the route.
+     *
+     * Used with the "route" option.
+     *
      * @var array
      */
     protected $routeParameters = array();
+
+    /**
+     * Set to false to not render
+     *
+     * @var boolean
+     */
+    protected $display = true;
+
+    /**
+     * Set to false to not render the children.
+     *
+     * @var boolean
+     */
+    protected $displayChildren = true;
+
+    /**
+     * Generate an absolute route
+     *
+     * To be used with the "content" or "route" option.
+     *
+     * @var boolean
+     */
+    protected $routeAbsolute = false;
 
     /**
      * @var boolean
@@ -289,7 +356,8 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     /**
      * Set the content document associated with this menu node
      *
-     * NOTE: Content documents must be mapped by PHPCR-ODM so that it can be persisted.
+     * NOTE: The content document must be mapped by PHPCR-ODM so that it can be
+     * persisted.
      *
      * @param object $content
      *
@@ -317,7 +385,10 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     }
 
     /**
-     * Specify if the content should be referenced weakly.
+     * Specify if the content should be referenced weakly or strongly.
+     *
+     * If there is existing referenced content, it is moved to become weak
+     * resp. hard referenced.
      *
      * @param $weak boolean
      *
@@ -364,10 +435,10 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     /**
      * Return the given attribute, optionally specifying a default value
      *
-     * @param  string $name     The name of the attribute to return
-     * @param  mixed  $default  The value to return if the attribute doesn't exist
+     * @param string $name     The name of the attribute to return
+     * @param string $default  The value to return if the attribute doesn't exist
      *
-     * @return mixed
+     * @return string
      */
     public function getAttribute($name, $default = null)
     {
@@ -381,8 +452,8 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     /**
      * Set the named attribute
      *
-     * @param $name string - attribute name
-     * @param $value mixed - attribute value
+     * @param string $name  attribute name
+     * @param string $value attribute value
      *
      * @return MenuNode - this instance
      */
@@ -406,7 +477,7 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     /**
      * Set the children attributes
      *
-     * @param $attributes array
+     * @param array $attributes
      *
      * @return MenuNode - this instance
      */
@@ -419,7 +490,7 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
 
     /**
      * Get all child menu nodes of this menu node. This will filter out all
-     * non-NodeInterface nodes.
+     * non-NodeInterface children.
      *
      * @return MenuNode[]
      */
@@ -437,7 +508,22 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     }
 
     /**
-     * Gets the route parameters
+     * Add a child menu node, automatically setting the parent node.
+     *
+     * @param MenuNode $child
+     *
+     * @return MenuNode - The newly added child node.
+     */
+    public function addChild(MenuNode $child)
+    {
+        $child->setParent($this);
+        $this->children[] = $child;
+
+        return $child;
+    }
+
+    /**
+     * Gets the route parameters.
      *
      * @return array
      */
@@ -447,13 +533,161 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
     }
 
     /**
-     * Sets the route parameters
+     * Sets the route parameters.
      *
-     * @param array the parameters
+     * @param array $routeParameters
+     *
+     * @return MenuNode - this instance
      */
     public function setRouteParameters($routeParameters)
     {
         $this->routeParameters = $routeParameters;
+
+        return $this;
+    }
+
+    /**
+     * Get extra information associated with this node.
+     *
+     * @return array
+     */
+    public function getExtras()
+    {
+        return $this->extras;
+    }
+
+    /**
+     * Set extra information associated with this node.
+     *
+     * @param array $extras
+     *
+     * @return MenuNode - this instance
+     */
+    public function setExtras(array $extras)
+    {
+        $this->extras = $extras;
+
+        return $this;
+    }
+
+    /**
+     * Get the link HTML attributes.
+     *
+     * @return array
+     */
+    public function getLinkAttributes()
+    {
+        return $this->linkAttributes;
+    }
+
+    /**
+     * Set the link HTML attributes as associative array.
+     *
+     * @param array $linkAttributes
+     *
+     * @return MenuNode - this instance
+     */
+    public function setLinkAttributes($linkAttributes)
+    {
+        $this->linkAttributes = $linkAttributes;
+
+        return $this;
+    }
+
+    /**
+     * Get the label HTML attributes.
+     *
+     * @return array
+     */
+    public function getLabelAttributes()
+    {
+        return $this->labelAttributes;
+    }
+
+    /**
+     * Set the label HTML attributes as associative array.
+     *
+     * @param array $labelAttributes
+     *
+     * @return MenuNode - this instance
+     */
+    public function setLabelAttributes($labelAttributes)
+    {
+        $this->labelAttributes = $labelAttributes;
+        return $this;
+    }
+
+    /**
+     * Whether to display this menu node.
+     *
+     * @return boolean
+     */
+    public function getDisplay()
+    {
+        return $this->display;
+    }
+
+    /**
+     * Set whether to display this menu node.
+     *
+     * @param boolean $display
+     *
+     * @return MenuNode - this instance
+     */
+    public function setDisplay($display)
+    {
+        $this->display = $display;
+
+        return $this;
+    }
+
+    /**
+     * Whether to display the children of this menu node.
+     *
+     * @return boolean
+     */
+    public function getDisplayChildren()
+    {
+        return $this->displayChildren;
+    }
+
+    /**
+     * Set whether to display the children of this menu node.
+     *
+     * @param boolean $displayChildren
+     *
+     * @return MenuNode - this instance
+     */
+    public function setDisplayChildren($displayChildren)
+    {
+        $this->displayChildren = $displayChildren;
+
+        return $this;
+    }
+
+    /**
+     * Whether to generate absolute links for route or content.
+     *
+     * @return boolean
+     */
+    public function getRouteAbsolute()
+    {
+        return $this->routeAbsolute;
+    }
+
+    /**
+     * Set whether to generate absolute links when generating from a route
+     * or the content.
+     *
+     * @param boolean $routeAbsolute
+     *
+     * @return MenuNode - this instance
+     */
+    public function setRouteAbsolute($routeAbsolute)
+    {
+        $this->routeAbsolute = $routeAbsolute;
+
+        return $this;
     }
 
     /**
@@ -467,88 +701,97 @@ class MenuNode implements NodeInterface, PublishTimePeriodWriteInterface, Publis
             'label' => $this->getLabel(),
             'attributes' => $this->getAttributes(),
             'childrenAttributes' => $this->getChildrenAttributes(),
-            'display' => true,
-            'displayChildren' => true,
+            'display' => $this->display,
+            'displayChildren' => $this->displayChildren,
             'content' => $this->getContent(),
             'routeParameters' => $this->getRouteParameters(),
-            // TODO provide the following information
-            'routeAbsolute' => false,
-            'linkAttributes' => array(),
-            'labelAttributes' => array(),
+            'routeAbsolute' => $this->routeAbsolute,
+            'linkAttributes' => $this->linkAttributes,
+            'labelAttributes' => $this->labelAttributes,
+
         );
-    }
-
-    /**
-     * Get extra attributes
-     *
-     * @return array
-     */
-    public function getExtras()
-    {
-        return $this->extras;
-    }
-
-    /**
-     * Set the extra attributes
-     *
-     * @param $extras array
-     *
-     * @return MenuNode - this instance
-     */
-    public function setExtras(array $extras)
-    {
-        $this->extras = $extras;
-
-        return $this;
-    }
-
-    /**
-     * Add a child menu node, automatically setting the parent node.
-     *
-     * @param MenuNode - Menu node to add
-     *
-     * @return MenuNode - The newly added child node.
-     */
-    public function addChild(MenuNode $child)
-    {
-        $child->setParent($this);
-        $this->children[] = $child;
-
-        return $child;
     }
 
     public function __toString()
     {
-        return $this->getLabel() ? : '(no label set)';
+        return $this->getLabel() ? (string) $this->getLabel() : '(no label set)';
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function isPublishable()
     {
         return $this->publishable;
     }
 
+    /**
+     * Set the publishable workflow flag.
+     *
+     * @param boolean $publishable
+     */
     public function setPublishable($publishable)
     {
         $this->publishable = $publishable;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function getPublishStartDate()
     {
         return $this->publishStartDate;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function setPublishStartDate(\DateTime $date = null)
     {
         $this->publishStartDate = $date;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function getPublishEndDate()
     {
         return $this->publishEndDate;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function setPublishEndDate(\DateTime $date = null)
     {
         $this->publishEndDate = $date;
+    }
+
+    /**
+     * Get the link type
+     *
+     * The link type is used to explicitly determine which of the uri, route 
+     * and content fields are used to determine the link which will bre 
+     * rendered for the menu item. If it is empty this will be determined 
+     * automatically.
+     *
+     * @return string
+     */
+    public function getLinkType() 
+    {
+        return $this->linkType;
+    }
+
+    /**
+     * @see getLinkType
+     * @see ContentAwareFactory::$validLinkTypes
+     *
+     * Valid link types are defined in ContenentAwareFactory
+     *
+     * @param $linkType string - one of uri, route or content
+     */
+    public function setLinkType($linkType)
+    {
+        $this->linkType = $linkType;
     }
 }