+++ Voter/UriPrefixVoter.php
@@ -7,9 +7,10 @@ use Symfony\Component\Routing\Route;
 use Knp\Menu\ItemInterface;
 
 /**
- * This voter checks if the content option is a Symfony Route instance and if
- * so compares its option "currentUriPrefix" with the request path. This allows
- * to configure a menu entry to be the current entry for a whole sub path.
+ * This voter checks if the content entry in the menu item extras is a Symfony
+ * Route instance and if so compares its option "currentUriPrefix" with the
+ * request path. This allows to configure a menu entry to be the current entry
+ * for a whole sub path.
  *
  * This voter is NOT enabled by default. Enable it in your bundle configuration
  * and set up a currentUriPrefix on menu content that is routes.
@@ -38,9 +39,10 @@ class UriPrefixVoter implements VoterInterface
             return null;
         }
 
-        $options = $item->getAttributes();
-        if ($options['content'] instanceof Route && $options['content']->hasOption('currentUriPrefix')) {
-            $currentUriPrefix = $options['content']->getOption('currentUriPrefix');
+        $content = $item->getExtra('content');
+
+        if ($content instanceof Route && $content->hasOption('currentUriPrefix')) {
+            $currentUriPrefix = $content->getOption('currentUriPrefix');
             $currentUriPrefix = str_replace('{_locale}', $this->request->getLocale(), $currentUriPrefix);
             if (0 === strncmp($this->request->getPathinfo(), $currentUriPrefix, strlen($currentUriPrefix))) {
                 return true;